requirements.txt
python-dotenv
streamlit
langchain-openai